feat(libsinsp): move resource utilization to sinsp stats
- Prepare m_sinsp_stats_v2 buffer to hold more stats and metrics than resource utilization metrics
- Expand resource utilization metrics (e.g. include overall host CPU and memory usages)
