+++ userspace/libsinsp/CMakeLists.txt
@@ -124,7 +124,6 @@ set(SINSP_SOURCES
 	threadinfo.cpp
 	tuples.cpp
 	sinsp.cpp
-	stats.cpp
 	token_bucket.cpp
 	stopwatch.cpp
 	utils.cpp
@@ -170,7 +169,7 @@ if(NOT MINIMAL_BUILD AND NOT EMSCRIPTEN)
 		list(APPEND SINSP_SOURCES
 			cgroup_limits.cpp
 			container_engine/cri.cpp
-			sinsp_resource_utilization.cpp
+			stats.cpp
 			grpc_channel_registry.cpp
 		)
 	endif()
